AI-teste/requirements.txt
# Runtime dependencies
pywebview>=5.0.0
google-genai>=1.5.0
pdfminer.six>=20221105
Pillow>=10.0.0

# Build / packaging (dev only)
pyinstaller>=6.0.0
